requirements-render.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
coverage==7.11.3
cryptography==46.0.3
decorator==4.4.2
dj-database-url==3.0.1
Django==5.2.7
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==37.12.0
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
instagrapi==2.2.1
moviepy==1.0.3
numpy==2.3.3
packaging==25.0
paypalrestsdk==1.13.3
pillow==11.3.0
pluggy==1.6.0
proglog==0.1.12
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.3.0
PySocks==1.7.1
pytest==9.0.0
pytest-cov==7.0.0
pytest-django==4.11.1
python-slugify==8.0.4
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==13.1.2
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
psycopg2-binary==2.9.9
django-channels==4.0.0
daphne==4.0.0